move azure client
