+++ e2e/config/config.go
@@ -4,7 +4,6 @@ import (
 	"os"
 	"strings"
 
-	"github.com/Azure/agentbakere2e/azure"
 	"github.com/Azure/agentbakere2e/toolkit"
 )
 
@@ -17,7 +16,7 @@ var (
 	ScenariosToRun     = envmap("SCENARIOS_TO_RUN")
 	ScenariosToExclude = envmap("SCENARIOS_TO_EXCLUDE")
 	KeepVMSS           = strings.EqualFold(os.Getenv("KEEP_VMSS"), "true")
-	Azure              = azure.MustNewAzureClient(mustenv("SUBSCRIPTION_ID"))
+	Azure              = MustNewAzureClient(mustenv("SUBSCRIPTION_ID"))
 )
 
 func mustenv(env string) string {
